clip/pyproject.toml
[tool.poetry]
name = "clip"
version = "0.1.0"
description = ""
authors = ["Teck Meng <furyx@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
open-clip-torch = "^2.26.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
